Factored code out of PathSolver class into new TestCaseBuilder class.
PathSolver was responsible for managing the constraints and underlying
z3 solver, but also for building test cases once a solution had been
found.  Splitting the code into two classes for these two tasks makes
work easier.
